# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AssignGuidance:
    properties:
      assignee_id:
        type: string
      guidance_template_id:
        type: string
    required:
    - assignee_id
    - guidance_template_id
    type: object
...
  dto.CreateIncidentDto:
    properties:
      alarm_id:
        type: string
      assignee_id:
        type: string
      description:
        type: string
      guidance_template_id:
        type: string
      location:
        type: string
      name:
        type: string
      severity:
        type: string
    required:
    - alarm_id
    - assignee_id
    - description
    - guidance_template_id
    - location
    - name
    - severity
    type: object
...
  dto.UpdateIncidentDto:
    properties:
      status:
        type: string
    required:
    - status
    type: object
...
  errors.ErrorDetail:
    properties:
      details: {}
      message:
        type: string
      type:
        $ref: '#/definitions/errors.ErrorType'
    type: object
  errors.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/errors.ErrorDetail'
      request_id:
        type: string
      status:
        type: integer
      timestamp:
        type: string
    type: object
  errors.ErrorType:
    enum:
    - VALIDATION_ERROR
    - NOT_FOUND
    - UNAUTHORIZED
    - FORBIDDEN
    - BAD_REQUEST
    - CONFLICT
    - INTERNAL_ERROR
    - DATABASE_ERROR
    - EXTERNAL_SERVICE_ERROR
    - TIMEOUT_ERROR
    type: string
    x-enum-varnames:
    - ErrorTypeValidation
    - ErrorTypeNotFound
    - ErrorTypeUnauthorized
    - ErrorTypeForbidden
    - ErrorTypeBadRequest
    - ErrorTypeConflict
    - ErrorTypeInternal
    - ErrorTypeDatabase
    - ErrorTypeExternal
    - ErrorTypeTimeout
  models.Alarm:
    properties:
      created_at:
        type: string
      description:
        type: string
      device:
        type: string
      id:
        type: string
      premise:
        $ref: '#/definitions/models.Premise'
      premise_id:
        type: string
      severity:
        type: string
      status:
        type: string
      triggered_at:
        type: string
      type:
        type: string
    type: object
  models.GuidanceStep:
    properties:
      created_at:
        type: string
      description:
        type: string
      guidance_template:
        $ref: '#/definitions/models.GuidanceTemplate'
      guidance_template_id:
        type: string
      id:
        type: string
      step_number:
        type: integer
      title:
        type: string
    type: object
  models.GuidanceTemplate:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      guidance_steps:
        items:
          $ref: '#/definitions/models.GuidanceStep'
        type: array
      id:
        type: string
      name:
        type: string
    type: object
  models.Incident:
    properties:
      alarm:
        $ref: '#/definitions/models.Alarm'
      alarm_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      incident_guidance:
        $ref: '#/definitions/models.IncidentGuidance'
      incident_media:
        items:
          $ref: '#/definitions/models.IncidentMedia'
        type: array
      location:
        type: string
      name:
        type: string
      severity:
        type: string
      status:
        type: string
    type: object
  models.IncidentGuidance:
    properties:
      assignee:
        $ref: '#/definitions/models.User'
      assignee_id:
        type: string
      assigner:
        $ref: '#/definitions/models.User'
      assigner_id:
        type: string
      created_at:
        type: string
      guidance_template:
        $ref: '#/definitions/models.GuidanceTemplate'
      guidance_template_id:
        type: string
      id:
        type: string
      incident:
        $ref: '#/definitions/models.Incident'
      incident_guidance_steps:
        items:
          $ref: '#/definitions/models.IncidentGuidanceStep'
        type: array
      incident_id:
        type: string
    type: object
  models.IncidentGuidanceStep:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      incident_guidance:
        $ref: '#/definitions/models.IncidentGuidance'
      incident_guidance_id:
        type: string
      is_completed:
        type: boolean
      step_number:
        type: integer
      title:
        type: string
    type: object
  models.IncidentMedia:
    properties:
      created_at:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      file_url:
        type: string
      id:
        type: string
      incident:
        $ref: '#/definitions/models.Incident'
      incident_id:
        type: string
      media_type:
        type: string
    type: object
  models.Premise:
    properties:
      address:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      parent_premise:
        $ref: '#/definitions/models.Premise'
      parent_premise_id:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  types.IncidentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Incident'
        type: array
      pagination:
        $ref: '#/definitions/types.Pagination'
    type: object
  types.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Smart City System (SCS) Operator API for managing premises, alarms,
    incidents, guidance templates, and guards
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SCS Operator API
  version: "1.0"
paths:
...
  /incidents:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all incidents
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.IncidentListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get incidents with pagination
      tags:
      - incidents
    post:
      consumes:
      - application/json
      description: Create a new incident with the provided information
      parameters:
      - description: Incident creation data
        in: body
        name: incident
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIncidentDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Incident'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new incident
      tags:
      - incidents
  /incidents/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific incident by its ID
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Incident'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get incident by ID
      tags:
      - incidents
    patch:
      consumes:
      - application/json
      description: Update an existing incident's status or other properties
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      - description: Incident update data
        in: body
        name: incident
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIncidentDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Incident'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update incident
      tags:
      - incidents
  /incidents/{id}/assign-guidance:
    post:
      consumes:
      - application/json
      description: Assign a guidance template to an incident with an assignee
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      - description: Guidance assignment data
        in: body
        name: guidance
        required: true
        schema:
          $ref: '#/definitions/dto.AssignGuidance'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IncidentGuidance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign guidance to incident
      tags:
      - incidents
  /incidents/{id}/complete:
    patch:
      consumes:
      - application/json
      description: Mark an incident as resolved/completed
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Complete incident
      tags:
      - incidents
  /incidents/{id}/guidance:
    get:
      consumes:
      - application/json
      description: Get the guidance assigned to a specific incident
      parameters:
      - description: Incident ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IncidentGuidance'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get incident guidance
      tags:
      - incidents
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"